Refactor into separate components
